Creates lint command
