+++ tools/cli/tsconfig.json
@@ -15,6 +15,9 @@
       "path": "../build"
     },
     {
+      "path": "../lint"
+    },
+    {
       "path": "../test"
     }
   ]
