+++ tools/lint/src/index.ts
@@ -1,96 +1,30 @@
 /* eslint-disable no-console */
-import chalk from 'chalk';
-import { spawn } from 'child_process';
-import { Command } from 'commander';
-import path from 'path';
-
 import { npmPkgJsonLint } from './npmPkgJsonLint';
 
-const rootDir = process.cwd();
-const eslintConfigPath = path.resolve(__dirname, '../config/eslint.config.js');
-const prettierConfigPath = path.resolve(
-  __dirname,
-  '../config/prettier.config.js',
-);
+import { eslint } from './eslint';
+import { LintCommandOptions } from './lint.types';
+import { prettier } from './prettier';
 
-const esLintExtensions = ['js', 'ts', 'tsx'];
-const prettierExtensions = [...esLintExtensions, 'mjs', 'json', 'md', 'yml'];
+export const lint = (options: LintCommandOptions) => {
+  const { fix, prettierOnly, eslintOnly, pkgJsonOnly, verbose } = options;
 
-interface LintFlags {
-  fix: boolean;
-  eslintOnly: boolean;
-  prettierOnly: boolean;
-  pkgJsonOnly: boolean;
-  verbose: boolean;
-}
+  const linters: Array<Promise<unknown>> = [];
 
-const cli = new Command()
-  .option('-f, --fix', 'fix linting errors', false)
-  .option('-e, --eslintOnly', 'run eslint only', false)
-  .option('-p, --prettierOnly', 'run prettier only', false)
-  .option('--pkgJsonOnly', 'run npmPackageJsonLint only', false)
-  .option('--verbose', 'verbose mode', false)
-  .parse(process.argv);
-
-const { fix, prettierOnly, eslintOnly, pkgJsonOnly, verbose } =
-  cli.opts() as LintFlags;
-
-// If prettierOnly or eslintOnly is true, run only that linter
-if (prettierOnly || eslintOnly || pkgJsonOnly) {
-  let lintPromise: Promise<unknown>;
-
-  if (eslintOnly) {
-    lintPromise = eslint();
-  } else if (prettierOnly) {
-    lintPromise = prettier();
-  } else {
-    lintPromise = npmPkgJsonLint(fix);
+  if (!prettierOnly && !pkgJsonOnly) {
+    linters.push(eslint({ fix, verbose }));
   }
-  lintPromise.then(() => process.exit(0));
-} else {
-  // Otherwise, run all linters
-  (async () => {
-    await eslint();
-    await prettier();
-    await npmPkgJsonLint(fix);
-  })();
-}
+  if (!eslintOnly && !pkgJsonOnly) {
+    linters.push(prettier({ fix, verbose }));
+  }
+  if (!prettierOnly && !eslintOnly) {
+    linters.push(npmPkgJsonLint({ fix, verbose }));
+  }
 
-/** Spawns an eslint job */
-function eslint() {
-  return new Promise<void>(resolve => {
-    console.log(chalk.blue('Running ESLint...'));
-    spawn(
-      'eslint',
-      [
-        '--config',
-        eslintConfigPath,
-        `${rootDir}/**/*.{${esLintExtensions.join(',')}}`,
-        fix ? '--fix' : '--no-fix',
-        verbose ? '' : '--quiet',
-      ],
-      {
-        stdio: 'inherit',
-      },
-    ).on('close', resolve);
-  });
-}
-
-/** Spawns a prettier job */
-function prettier() {
-  return new Promise<void>(resolve => {
-    console.log(chalk.magenta('Running Prettier...'));
-    spawn(
-      'prettier',
-      [
-        fix ? '--write' : '--check',
-        '--config',
-        prettierConfigPath,
-        `${rootDir}/**/*.{${prettierExtensions.join(',')}}`,
-      ],
-      {
-        stdio: 'inherit',
-      },
-    ).on('close', resolve);
-  });
-}
+  Promise.all(linters)
+    .then(() => {
+      process.exit(0);
+    })
+    .catch(() => {
+      process.exit(1);
+    });
+};
