+++ tools/lint/src/lint.types.ts
@@ -1,0 +1,7 @@
+export interface LintCommandOptions {
+  fix: boolean;
+  eslintOnly: boolean;
+  prettierOnly: boolean;
+  pkgJsonOnly: boolean;
+  verbose: boolean;
+}
