+++ tools/lint/src/npmPkgJsonLint.ts
@@ -1,29 +1,37 @@
-/* eslint-disable no-console */
+import { spawn } from 'child_process';
 import chalk from 'chalk';
-import fs from 'fs';
-import {
-  NpmPackageJsonLint,
-  PackageJsonFileLintingResult,
-} from 'npm-package-json-lint';
 import path from 'path';
-
+import { LintCommandOptions } from './lint.types';
+// import { NpmPackageJsonLint } from 'npm-package-json-lint';
 const rootDir = process.cwd();
 const npmPkgLintConfigPath = path.resolve(
   __dirname,
   '../config/npmpackagejsonlintrc.config.js',
 );
 
-const npmPackageJsonLinter = new NpmPackageJsonLint({
-  cwd: rootDir,
-  patterns: ['**/package.json'],
-  configFile: npmPkgLintConfigPath,
-});
+// const npmPackageJsonLinter = new NpmPackageJsonLint({
+//   cwd: rootDir,
+//   patterns: ['**/package.json'],
+//   configFile: npmPkgLintConfigPath,
+// });
 
 /** Spawns a npmPkgJsonLint job */
-export function npmPkgJsonLint(fix: boolean) {
+export function npmPkgJsonLint({
+  fix,
+  verbose,
+}: Pick<LintCommandOptions, 'fix' | 'verbose'>) {
   return new Promise<void>((resolve, reject) => {
     console.log(chalk.yellow('Running npmPkgJsonLint...'));
 
+    spawn('npmPkgJsonLint', [rootDir, '--configFile', npmPkgLintConfigPath], {
+      cwd: rootDir,
+      stdio: 'inherit',
+    })
+      .on('close', resolve)
+      .on('error', reject);
+
+    /*
+    TODO: use the JS API
     const { results, errorCount } = npmPackageJsonLinter.lint();
 
     results.forEach(result => {
@@ -41,5 +49,6 @@ export function npmPkgJsonLint(fix: boolean) {
     } else {
       resolve();
     }
+    */
   });
 }
