+++ tools/lint/tsconfig.json
@@ -1,3 +1,13 @@
 {
   "extends": "@lg-tools/build/config/package.tsconfig.json",
+  "compilerOptions": {
+    "declarationDir": "dist",
+    "outDir": "dist",
+    "rootDir": "src",
+    "baseUrl": "."
+  },
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": ["../../node_modules/**/*"]
 }
